feat: Add JD comment automation modules
- Implemented `comment_generator.py` for generating product reviews based on product names using templates and keyword extraction.
- Created `config.py` to manage configuration constants and file loading logic for the JD review tool.
- Developed `evaluator.py` for querying and summarizing review statistics from JD.
- Added `handlers.py` to process regular reviews, follow-up reviews, and service ratings, ensuring all HTTP requests use explicit headers.
- Introduced `image_utils.py` for downloading, uploading, and managing images during the review process.
- Established a logging system in `logger.py` with ANSI color formatting for better readability.
- Defined review templates in `templates.py` for default and gift comments.
