+++ jdspider.py
@@ -16,29 +16,10 @@ import yaml
 import zhon.hanzi
 from lxml import etree
 
+from templates import DEFAULT_COMMENTS
+
 # 配置日志（延迟初始化，由外部设置）
 default_logger = logging.getLogger("jdspider")
-
-# 默认评价模板（与 auto_comment_plus.py 保持一致）
-DEFAULT_COMMENTS = [
-    "考虑买这个$之前我是有担心过的，因为我不知道$的质量和品质怎么样，但是看了评论后我就放心了。",
-    "买这个$之前我是有看过好几家店，最后看到这家店的评价不错就决定在这家店买 ",
-    "看了好几家店，也对比了好几家店，最后发现还是这一家的$评价最好。",
-    "看来看去最后还是选择了这家。",
-    "之前在这家店也买过其他东西，感觉不错，这次又来啦。",
-    "这家的$的真是太好用了，用了第一次就还想再用一次。",
-    "收到货后我非常的开心，因为$的质量和品质真的非常的好！",
-    "拆开包装后惊艳到我了，这就是我想要的$!",
-    "快递超快！包装的很好！！很喜欢！！！",
-    "包装的很精美！$的质量和品质非常不错！",
-    "收到快递后迫不及待的拆了包装。$我真的是非常喜欢",
-    "真是一次难忘的购物，这辈子没见过这么好用的东西！！",
-    "经过了这次愉快的购物，我决定如果下次我还要买$的话，我一定会再来这家店买的。",
-    "不错不错！",
-    "我会推荐想买$的朋友也来这家店里买",
-    "真是一次愉快的购物！",
-    "大大的好评!以后买$再来你们店！(￣▽￣)",
-]
 
 # Cookie（由外部设置）
 cookie: bytes = b""
