+++ templates.py
@@ -1,0 +1,45 @@
+# -*- coding: utf-8 -*-
+"""评论模板模块。
+
+本模块是 DEFAULT_COMMENTS 和 GIFT_COMMENTS 的唯一定义来源，
+其他模块必须从此处导入，不允许重复定义。
+"""
+
+from __future__ import annotations
+
+# 默认评价模板
+DEFAULT_COMMENTS: list[str] = [
+    "考虑买这个$之前我是有担心过的，因为我不知道$的质量和品质怎么样，但是看了评论后我就放心了。",
+    "买这个$之前我是有看过好几家店，最后看到这家店的评价不错就决定在这家店买 ",
+    "看了好几家店，也对比了好几家店，最后发现还是这一家的$评价最好。",
+    "看来看去最后还是选择了这家。",
+    "之前在这家店也买过其他东西，感觉不错，这次又来啦。",
+    "这家的$的真是太好用了，用了第一次就还想再用一次。",
+    "收到货后我非常的开心，因为$的质量和品质真的非常的好！",
+    "拆开包装后惊艳到我了，这就是我想要的$!",
+    "快递超快！包装的很好！！很喜欢！！！",
+    "包装的很精美！$的质量和品质非常不错！",
+    "收到快递后迫不及待的拆了包装。$我真的是非常喜欢",
+    "真是一次难忘的购物，这辈子没见过这么好用的东西！！",
+    "经过了这次愉快的购物，我决定如果下次我还要买$的话，我一定会再来这家店买的。",
+    "不错不错！",
+    "我会推荐想买$的朋友也来这家店里买",
+    "真是一次愉快的购物！",
+    "大大的好评!以后买$再来你们店！(￣▽￣)",
+]
+
+# 赠品评价模板
+GIFT_COMMENTS: list[str] = [
+    "赠品挺好的。",
+    "很贴心，能有这样免费赠送的赠品!",
+    "正好想着要不要多买一份增值服务，没想到还有这样的赠品。",
+    "赠品正合我意。",
+    "赠品很好，挺不错的。",
+    "本来买了产品以后还有些担心。但是看到赠品以后就放心了。",
+    "不论品质如何，至少说明店家对客的态度很好！",
+    "我很喜欢这些商品！",
+    "我对于商品的附加值很在乎，恰好这些赠品为这件商品提供了这样的的附加值，这令我很满意。",
+    "感觉现在的网购环境环境越来越好了，以前网购的时候还没有过么多贴心的赠品和增值服务",
+    "第一次用京东，被这种赠品和增值服物的良好态度感动到了。",
+    "赠品还行。",
+]
